Changes MalformedExpressionErrors to isError
Deciding that the client would need to look for
MalformedExpressionErrors did not work, because in that case a user
would have two different possible error conditions: a manipulation
returns an error that responds to isError() with true, or
a manipulation that throws an exception.

Instead, methods are reworked s.t. the only error condition
is the isError() == true condition.